Move Exceptions to service.Exceptions
