+++ src/main/java/com/simbirsoft/shatrov/WarehouseApp/controller/UserController.java
@@ -1,11 +1,10 @@
 package com.simbirsoft.shatrov.WarehouseApp.controller;
 
-import com.simbirsoft.shatrov.WarehouseApp.Exceptions.*;
+import com.simbirsoft.shatrov.WarehouseApp.service.Exceptions.*;
 import com.simbirsoft.shatrov.WarehouseApp.entity.User;
 import com.simbirsoft.shatrov.WarehouseApp.service.user.UserService;
 import lombok.SneakyThrows;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
